requirements.txt
git+https://github.com/huggingface/transformers
datasets
loralib
sentencepiece
bitsandbytes
accelerate
langchain
einops
faiss-cpu
langchain_community
chromadb
bs4
qdrant-client
langchainhub
langchain-huggingface
langchain-cohere
wikipedia
arxiv
pymupdf
xmltodict
cohere
sentence_transformers
torch
ragas
python-dotenv
langchain-qdrant